Score UI
Created a basic design of the score ui which will be used to keep track of the score and announce who wins after each round.

Note: I also changed the names from "Player" and "Computer" to "Player1" and "Player2" due to the length of the words. I wanted the words to be evenly lengthed.